Dhriti_ai-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
alembic
email-validator
pandas
openpyxl
python-multipart
